Split index into multiple files
